Split server tests into per-category files
Move from monolithic test_server.rs to 15 focused test files + shared
common/mod.rs module. Same 50 tests, same coverage, better organization:

  tests/common/mod.rs      Shared infrastructure (servers, helpers, specs)
  tests/test_health.rs     Health endpoints (3 tests)
  tests/test_themes.rs     Theme endpoints (3 tests)
  tests/test_openapi.rs    OpenAPI spec + Swagger UI (2 tests)
  tests/test_vegalite.rs   Vega-Lite conversions (7 tests)
  tests/test_vega.rs       Vega conversions (6 tests)
  tests/test_svg.rs        SVG conversions (3 tests)
  tests/test_logs.rs       X-VLC-Logs header (2 tests)
  tests/test_request_id.rs X-Request-Id (2 tests)
  tests/test_validation.rs Request validation (3 tests)
  tests/test_overrides.rs  Spec overrides (1 test)
  tests/test_auth.rs       Bearer token auth (6 tests)
  tests/test_user_agent.rs User-Agent requirement (4 tests)
  tests/test_opaque.rs     Opaque errors (1 test)
  tests/test_budget.rs     Budget tracking + admin (4 tests)
  tests/test_cors.rs       CORS (3 tests)
