+++ vl-convert-server/tests/test_svg.rs
@@ -1,0 +1,49 @@
+mod common;
+
+use common::*;
+use serde_json;
+
+#[tokio::test]
+async fn test_svg_to_png() {
+    let server = &*DEFAULT_SERVER;
+    let resp = server
+        .client
+        .post(format!("{}/svg/png", server.base_url))
+        .json(&serde_json::json!({"svg": simple_svg()}))
+        .send()
+        .await
+        .unwrap();
+    assert_eq!(resp.status(), 200);
+    let bytes = resp.bytes().await.unwrap();
+    assert_eq!(&bytes[..4], &[0x89, b'P', b'N', b'G'], "bad PNG magic");
+}
+
+#[tokio::test]
+async fn test_svg_to_jpeg() {
+    let server = &*DEFAULT_SERVER;
+    let resp = server
+        .client
+        .post(format!("{}/svg/jpeg", server.base_url))
+        .json(&serde_json::json!({"svg": simple_svg()}))
+        .send()
+        .await
+        .unwrap();
+    assert_eq!(resp.status(), 200);
+    let bytes = resp.bytes().await.unwrap();
+    assert_eq!(&bytes[..2], &[0xFF, 0xD8], "bad JPEG magic");
+}
+
+#[tokio::test]
+async fn test_svg_to_pdf() {
+    let server = &*DEFAULT_SERVER;
+    let resp = server
+        .client
+        .post(format!("{}/svg/pdf", server.base_url))
+        .json(&serde_json::json!({"svg": simple_svg()}))
+        .send()
+        .await
+        .unwrap();
+    assert_eq!(resp.status(), 200);
+    let bytes = resp.bytes().await.unwrap();
+    assert!(bytes.starts_with(b"%PDF"), "expected PDF magic");
+}
